requirements.txt
numpy>=1.24
opencv-python>=4.8
PyYAML>=6.0
matplotlib>=3.7
scikit-image>=0.22
scipy>=1.10
ultralytics>=8.2
